clean up append output
