+++ tests/unit/utils/test_output.py
@@ -20,7 +20,7 @@ configs_to_test = [dict(), minimal_config]
 
 class TestInit:
     def test_empty_init(self, root):
-        output = Output(root=root)
+        output = Output(root=root, run_name="test")
         print(output.root)
         print(output.workdir)
         assert isdir(output.root)
